refactor(platform): split extensions/defaults.py into job-scoped sub-modules behind the facade
The 1,815-line module carried four jobs: extension-seam policy defaults,
the ProcessingPort, the CatalogPort, and the AI providers. Move each class
verbatim into defaults_extensions.py, defaults_processing_port.py,
defaults_catalog_port.py, defaults_ai_anthropic.py, and
defaults_ai_openai.py, and keep defaults.py as the import facade
(all names in __all__, mirroring the datasets domain/service.py
convention) so no import site churns.

Path-anchored guards follow the move: the AUDIT-02 log_action git-grep
exclusion and the C901 burn-down entry retarget to the sub-modules that
now hold DefaultAuditSink.emit() and the OpenAI complete() loop.

Part of #836.
